pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-rag"
version = "0.1.0"
description = "Obsidian Vault RAG skill for OpenClaw"
requires-python = ">=3.9,<3.13"
dependencies = [
    "python-frontmatter>=1.1.0",
    "fastembed==0.7.4",
    "lancedb==0.25.3",
    "click>=8.1.7",
    "onnxruntime==1.19.2",
    "platformdirs>=2.5.0",
]

[project.scripts]
obsidian-rag = "obsidian_rag.__main__:cli"
